% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleComp.R
\name{AlleleComp}
\alias{AlleleComp}
\title{Copynumber estimation}
\usage{
AlleleComp(data, AD, file.name, method, uniform.break)
}
\arguments{
\item{data}{A \code{vcfR} object of the sequencing calls.}

\item{AD}{a \code{character} deoning \emph{ID} for depth of the reference allele.
This is often separately present in the VCF file. Default is \code{NULL}.}

\item{file.name}{an optional \code{character} to define output file name. Default is \emph{tumor.sample}.}

\item{method}{Algorithm to be used for copy number calculations. options include "apriori" wich
uses \code{\link{CopySeg_sequenza}} and "naive" using \code{\link{CopySeg_falcon}}.}

\item{uniform.break}{A numeric value signifying fixed length of the genomic window. Each window
is considered as distinct chromosomal segment with edges being the break points for copy number
estimation. A good window length is 1Mb (i.e. 1e6)}
}
\value{
A transformed \code{dataframe} usable in \emph{CloneStrat} that represents data on all variants
in the .vcf file. It returns summaries on the variants with the collumn \emph{CN.profile} depicting
the estimated allelic compositions.
}
\description{
Allelic segmentations are estimated for one sample at a time with unfiltered sequencing calls.
}
\details{
The function writes a \emph{.txt} data in working directory with the name defined in \code{file.name} used by \emph{sequenza}.
The output file written can be used in conjunction with post variants call sequence file. These can be merged and used for surther analysis
with \code{\link{cluster.doc}} or \code{\link{seqn.scale}}
}
\examples{
\donttest{#AlleleComp(data = x, AD = "AD", method = "naive")}

}
\seealso{
\code{\link{segment.plot}}
}
