% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.doc.R
\name{cluster.doc}
\alias{cluster.doc}
\title{Clonal deconvolution}
\usage{
cluster.doc(
  data = NULL,
  sample = NULL,
  vaf = NULL,
  allele.comp = NULL,
  n.clone = NULL,
  n.subclone = NULL,
  optimization.method = "GMM",
  clustering.method = "HKM",
  clonality = "Allelic composition",
  instruct = TRUE
)
}
\arguments{
\item{data}{A \code{dataframe} containing summary from DNA sequencing. It must include a column of sample IDs
and a corresponding column with the variant allele frequencies.}

\item{sample}{\code{Integer or character} of the column name or column number of the sample IDs.}

\item{vaf}{\code{Integer or character} of the column name or column number of the variant allele frequency.}

\item{allele.comp}{\code{Character} string for allelic composition of the variants. example: '1+1' or '2+3' etc.}

\item{n.clone}{Optional \code{integer} for number of suspected clones, default NULL.}

\item{n.subclone}{Optional \code{interger} for number of suspected subclones, default NULL.}

\item{optimization.method}{Method to find optimal number of clusters; \emph{GMM} or \emph{bootstrap}. Default is \emph{GMM}.}

\item{clustering.method}{Clustering methods; \emph{HKM}, \emph{bootkm} or \emph{hybrid}. Default is \emph{hkm}.}

\item{clonality}{Method for determining clonality of the predicted clusters; \emph{Allelic composition} (default) or \emph{density}}

\item{instruct}{\code{Character} input for accepting program suggestion.}
}
\value{
A list of 12 objects is returned that includes all the summary statistics, diagnositics and the predictions as well as the mapping
internally used for clonal deconvolution.

\code{predicted.data} is necessarily an extension to the input \code{data} with the addition of the predicted clone and sub-clone
status of each variant for corresponding samples.

\code{density.map} is a distance matrix convoluted from cluster distances and desity departures.

\code{collapse} are clusters that are initially prredicted but later collapsed on each other dues to similarity between them.

\code{fitted.hkm, fitted bootkm or fitted.hybrid} is a vector of initial cluster assignment by the algorithm chosen.
Only one of these will have an output and the rest will show \code{NA}.

\code{Number of unscaled clusters} gives umber of predicted clusters before collapsing with density estimates.

\code{Number of scaled clusters} gives number of predicted clusters after collapsing (if any).

\code{cluster.diagnostics} if the optimization method was chosen to be \emph{GMM}, this is an object of \code{S3} class
that includes clustering diagnostics from the model-based clustering. If the chosen method was \emph{bootstrap} then this is a list.

\code{cluster centers} are the centroids of the predicted scaled clusters.

\code{cluster mapping} provides the map between scaled clusters and the clonal deconvolution assignments

\code{Dunn index} is the Dunn index for the fitted cluster.
}
\description{
Clone / Sub-clone decomposition of DNA sequencing data. This is recommended to be used for more than one sample preferably collected from
the same individual at different times. If the sample qualities vary, it is recommended to perform scaling first with \code{\link{seqn.scale}}.
}
\details{
\code{cluster.doc} is meant to do two things, first determine the optimum number of clusters that \emph{should} be fitted and
second, to infer what groups the clusters thus obtained should be assigned to.

The data inputs interactively requested from the user help obtain the following information

\code{chromosomal segmentation} helps in determining the number of clone/sub-clone cloud to be expected in the data. As
variant alleles from different aberrant chromosomes may have similar relative frequencies but discordant clonal
interpretation. On the contrary convergent clonal alleles may demonstrate divergent frequencies if arisen from dissimilar aneuploidy.

\code{clouds} give the program a visual feedback from the user that assume to carry some biological interpretation of the
frequency distributions present in the data. This is a subjective estimate that the program later uses for cluster assignment.

Out of the two methods used for cluster optimization, \emph{GMM} stands for \emph{Gaussian Mixed Models} whereas \emph{bootstrap},
as the name suggests perform \emph{bootstrap} resampling of the VAFs in 50 repetitions with 20 runs each to find the most stable parameter
for clustering. \emph{GMM} outputs the optimization curve with \code{BIC} and \emph{AIC} against number of clusters chosen in the
\code{X-axis} where \emph{bootstrap} shows the \code{Smin} statistics instead in the \code{Y-axis}. Where as \code{gap}
calculates the gap statistics for each clustering. In all cases the statistics are to be interpreted as proxies for the \emph{entropy} of the
system. The maximum entropy is likely to indicate the most stable solution.

\code{clustering.method} gives the user three choices:

\code{HKM} is \emph{Heierarchical K-means clustering} which uses heierarchical clustering first to determine the cluster centers
that are subsequently used as the starting point for the K-means clustering.
\code{bootkm} performs a \emph{bootstrap} resampling of 20 fitted K-means clusters with 50 resamplings to out put the clusters.
\code{hybrid} performs \emph{hkm} on the principal component of the data.

\code{clonality} provides two choices for clonality assignment. The default is \emph{Allelic composition} that measures expected
clonality patterns according to the copy numbers. But in cases of unreliable allelic composition estimates this method may fail. In such
situations the clonality can be assigned without apriori assumptions with the alternate \emph{density} based method.
}
\examples{
\donttest{#cluster.doc(test.dat, 1, 2, optimization.method = 'GMM', clustering.method = 'HKM')}
}
\seealso{
\code{\link{seqn.scale}} \code{\link{cluster.doubt}}
}
