% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.doubt.R
\name{cluster.doubt}
\alias{cluster.doubt}
\title{User overriden clonal deconvolution}
\usage{
cluster.doubt(CD.obj, sample, vaf, sample.name, cluster.num)
}
\arguments{
\item{CD.obj}{A \code{cluster.doc} object}

\item{sample}{Collumn number of the \emph{predicted.data} from the \code{cluster.doc}
output that contain sample IDs}

\item{vaf}{Collumn number of the \emph{predicted.data} from the \code{cluster.doc}
output that contain variant allele frequencies used for the analysis.}

\item{sample.name}{a vector of sample IDs}

\item{cluster.num}{a numeric vetor of clone/sub-clonal split of respective sample}
}
\value{
A list of 3 objects

\code{fitted.cluster} includes the clustering results from the final fit with user input

\code{predicted.data} A dataframe shows the changed clustering results due to the user defined
clone / sub-clone smear for the selected samples
}
\description{
Sample specific user curated Clone / Sub-clone decomposition of DNA sequencing data
}
\examples{
\donttest{#cd.res<-cluster.doc(test.dat)
#cd.new<-cluster.doubt(cd.res,sample,vaf,c("Sample_1","Sample_3"),c(2,2,3,2))}
}
\seealso{
\code{\link{cluster.doc}}
}
