% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{Neuroblastoma}
\alias{Neuroblastoma}
\title{Human neuroblastoma data}
\format{
An object of class \code{"dataframe"}
}
\usage{
data(Neuroblastoma)
}
\value{
\code{Sample} is column of IDs corresponding to 4 samples (2x primary and 2x metastasis).

\code{VAF} denotes the variant allele frequencies.

\code{RefseqID} annotates each of the variants.
}
\description{
Exome sequencing data of human neuroblastoma tumor samples available in public library.
}
\examples{
\donttest{data(Neuroblastoma)}
}
\seealso{
\code{\link{primary_1}} \code{\link{primary_2}} \code{\link{metastasis_1}} \code{\link{metastasis_2}}

\href{https://doi.org/10.1038/s41588-018-0131-y}{Karlsson \emph{et al}., 2018}
}
\keyword{datasets}
