% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.auto.plot.R
\name{variant.auto.plot}
\alias{variant.auto.plot}
\title{Automated Multi-sample plot}
\usage{
variant.auto.plot(CD.obj, annotation.col)
}
\arguments{
\item{CD.obj}{A \code{cluster.doc object}}

\item{annotation.col}{name of the column containing annotations of the variants
in original WES \code{dataframe} used in the clonal deconvolution using \code{cluster.doc}}
}
\value{
Plot objects with the relevant annotation highlighted.

This function plots all variants present in the sample. Depending on the number of variants this can generate a \emph{lot}
of plots. All of these plots will be saved under a new directory named \code{img} inside the working directory.
Hence, it is important to check that there are no directory named \code{img} inside the working directory
}
\description{
Automated plotting of all variants present in the WES data
}
\examples{
\donttest{#cd.res<-cluster.doc(test.dat,1,2)
#variant.auto.plot(cd.res,'annotation')}
}
